jira/story-extractor/requirements.txt
jira>=3.6.0
pyyaml>=6.0
python-dotenv>=1.0.0
